backend/pyproject.toml
[project]
name = "protocol-ai-backend"
version = "0.1.0"
description = "AI-Protocolist backend (ASR + diarization + sign language)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "gunicorn>=23.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.17",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "email-validator>=2.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
    "alembic>=1.14",
    "redis>=5.2",
    "celery[redis]>=5.4",
    "flower>=2.0",
    "boto3>=1.35",
    "aioboto3>=13.2",
    "httpx>=0.27",
    "structlog>=24.4",
    "python-magic>=0.4; platform_system != 'Windows'",
    "python-magic-bin>=0.4; platform_system == 'Windows'",
    "slowapi>=0.1.9",
    "tenacity>=9.0",
    "orjson>=3.10",
    # ML (heavy — actually used by workers)
    "faster-whisper>=1.0",
    "pyannote.audio>=3.3",
    "torch>=2.4",
    "torchaudio>=2.4",
    "numpy>=1.26",
    "soundfile>=0.12",
    "librosa>=0.10",
    "ffmpeg-python>=0.2",
    "openai>=1.54",
    # Export
    "weasyprint>=63.0",
    "python-docx>=1.1",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-python-jose",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
strict_optional = true
warn_unused_ignores = true
